// Erp-Efatura-Dashboard/src/app/models/integration.ts
export class Integration {
    password: string;
    email: string;
    code: string;
    partyId: string;
    partyName: string;
    firstName: string;
    middleName: string;
    lastName: string;
    accountCodeList: string;
    pk: string;
    gb: string;
    certificateOwnerSerialNumber: string;
    certificateSerialNumber: string;
    pin: string;
    constructor() {
        this.partyId = '';
    }
}